Submodules as jj workspaces
jj workspaces allows to have several repositories sharing one same
backend.

This is typically something I would want to assure all the work I do on
one submodule is immediately accesible to the other. Not requiring my to
push. Which removes a lot of potential issues where I don't remember
where I worked.

So this implies a lot of changes, we need to trick git into knowing
it has control on its submodules while it does not. As jj workspaces
cannot be colocated and having git changing the content of the
submodule, managed by jj, just creates in jj a commit which all the
changes. So conflicts conflicts

Ideas:
- Extract from git:
  - remote URL (.gitmodules)
  - path (.gitmodules)
  - current commit the repository should be at (?)
- jj clone the root repository in the workspace (or use the one already
  there)
- jj --repository <root_repo> workspace add <path>

Required changes:
- "git status" for submodules cannot rely on git2:
  - extract the current commit, look at the actual commit if it is the
    same, then Ok else commit changed.
  - Since there is no stage in jj, then this would be the only
    information to handle.
- Updating submodule: The git commit operation for a submodule needs to be
fully re-implemented, the operations could be:
  - jj new
  - do the operation to git add submodule, without changing the
    submodule content.
  - git commit (since it cannot be done with jj)
  - jj snapshot